Refactor: extract manifest.go, add parseAndValidate helper
- Split slack.go (587 lines) by extracting Manifest() into manifest.go.
  slack.go now focuses on connector setup, HTTP client logic, and shared
  helpers (~250 lines). manifest.go owns all action schemas and templates.
- Add parseAndValidate() helper that replaces the 7-line
  json.Unmarshal + validate boilerplate repeated across all 10 action
  files. Each action now uses a 3-line call instead.
- Add validate() method to listChannelsParams for consistency with all
  other action param structs (was previously inline).
- Remove unused encoding/json and fmt imports from 8 action files.
