+++ connectors/slack/manifest.go
@@ -1,0 +1,350 @@
+package slack
+
+import (
+	"encoding/json"
+
+	"github.com/supersuit-tech/permission-slip-web/connectors"
+)
+
+// Manifest returns the connector's metadata manifest. Used by the server to
+// auto-seed DB rows on startup, replacing manual seed.go files.
+func (c *SlackConnector) Manifest() *connectors.ConnectorManifest {
+	return &connectors.ConnectorManifest{
+		ID:          "slack",
+		Name:        "Slack",
+		Description: "Slack integration for team communication",
+		Actions: []connectors.ManifestAction{
+			{
+				ActionType:  "slack.send_message",
+				Name:        "Send Message",
+				Description: "Send a message to a Slack channel",
+				RiskLevel:   "low",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"required": ["channel", "message"],
+					"properties": {
+						"channel": {
+							"type": "string",
+							"description": "Channel name (e.g. #general) or ID (e.g. C01234567)"
+						},
+						"message": {
+							"type": "string",
+							"description": "Message text (supports Slack mrkdwn formatting)"
+						}
+					}
+				}`)),
+			},
+			{
+				ActionType:  "slack.create_channel",
+				Name:        "Create Channel",
+				Description: "Create a new Slack channel",
+				RiskLevel:   "medium",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"required": ["name"],
+					"properties": {
+						"name": {
+							"type": "string",
+							"description": "Channel name (lowercase, no spaces, max 80 chars)"
+						},
+						"is_private": {
+							"type": "boolean",
+							"default": false,
+							"description": "Create as a private channel"
+						}
+					}
+				}`)),
+			},
+			{
+				ActionType:  "slack.list_channels",
+				Name:        "List Channels",
+				Description: "List Slack channels visible to the bot",
+				RiskLevel:   "low",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"properties": {
+						"types": {
+							"type": "string",
+							"default": "public_channel",
+							"description": "Comma-separated channel types: public_channel, private_channel, mpim, im"
+						},
+						"limit": {
+							"type": "integer",
+							"default": 100,
+							"description": "Max channels to return (1-1000)"
+						},
+						"cursor": {
+							"type": "string",
+							"description": "Pagination cursor from a previous response"
+						},
+						"exclude_archived": {
+							"type": "boolean",
+							"default": true,
+							"description": "Exclude archived channels from results"
+						}
+					}
+				}`)),
+			},
+			{
+				ActionType:  "slack.read_channel_messages",
+				Name:        "Read Channel Messages",
+				Description: "Read recent messages from a Slack channel",
+				RiskLevel:   "low",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"required": ["channel"],
+					"properties": {
+						"channel": {
+							"type": "string",
+							"description": "Channel ID (e.g. C01234567)"
+						},
+						"limit": {
+							"type": "integer",
+							"default": 20,
+							"description": "Max messages to return (1-1000)"
+						},
+						"oldest": {
+							"type": "string",
+							"description": "Only messages after this Unix timestamp"
+						},
+						"latest": {
+							"type": "string",
+							"description": "Only messages before this Unix timestamp"
+						},
+						"cursor": {
+							"type": "string",
+							"description": "Pagination cursor from a previous response"
+						}
+					}
+				}`)),
+			},
+			{
+				ActionType:  "slack.read_thread",
+				Name:        "Read Thread",
+				Description: "Read replies in a Slack thread",
+				RiskLevel:   "low",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"required": ["channel", "thread_ts"],
+					"properties": {
+						"channel": {
+							"type": "string",
+							"description": "Channel ID containing the thread (e.g. C01234567)"
+						},
+						"thread_ts": {
+							"type": "string",
+							"description": "Timestamp of the parent message (e.g. 1234567890.123456)"
+						},
+						"limit": {
+							"type": "integer",
+							"default": 50,
+							"description": "Max replies to return (1-1000)"
+						},
+						"cursor": {
+							"type": "string",
+							"description": "Pagination cursor from a previous response"
+						}
+					}
+				}`)),
+			},
+			{
+				ActionType:  "slack.schedule_message",
+				Name:        "Schedule Message",
+				Description: "Schedule a message for future delivery to a Slack channel",
+				RiskLevel:   "low",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"required": ["channel", "message", "post_at"],
+					"properties": {
+						"channel": {
+							"type": "string",
+							"description": "Channel name (e.g. #general) or ID (e.g. C01234567)"
+						},
+						"message": {
+							"type": "string",
+							"description": "Message text (supports Slack mrkdwn formatting)"
+						},
+						"post_at": {
+							"type": "integer",
+							"description": "Unix timestamp for when the message should be sent (must be in the future)"
+						}
+					}
+				}`)),
+			},
+			{
+				ActionType:  "slack.set_topic",
+				Name:        "Set Topic",
+				Description: "Update a Slack channel's topic",
+				RiskLevel:   "medium",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"required": ["channel", "topic"],
+					"properties": {
+						"channel": {
+							"type": "string",
+							"description": "Channel ID (e.g. C01234567)"
+						},
+						"topic": {
+							"type": "string",
+							"description": "New channel topic (max 250 characters)"
+						}
+					}
+				}`)),
+			},
+			{
+				ActionType:  "slack.invite_to_channel",
+				Name:        "Invite to Channel",
+				Description: "Invite one or more users to a Slack channel",
+				RiskLevel:   "medium",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"required": ["channel", "users"],
+					"properties": {
+						"channel": {
+							"type": "string",
+							"description": "Channel ID (e.g. C01234567)"
+						},
+						"users": {
+							"type": "string",
+							"description": "Comma-separated list of user IDs to invite (e.g. U01234567,U09876543)"
+						}
+					}
+				}`)),
+			},
+			{
+				ActionType:  "slack.upload_file",
+				Name:        "Upload File",
+				Description: "Upload a file to a Slack channel",
+				RiskLevel:   "low",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"required": ["channel", "filename", "content"],
+					"properties": {
+						"channel": {
+							"type": "string",
+							"description": "Channel ID (e.g. C01234567)"
+						},
+						"filename": {
+							"type": "string",
+							"description": "Name of the file (e.g. report.csv)"
+						},
+						"content": {
+							"type": "string",
+							"description": "File content as text"
+						},
+						"title": {
+							"type": "string",
+							"description": "Display title for the file (defaults to filename)"
+						}
+					}
+				}`)),
+			},
+			{
+				ActionType:  "slack.add_reaction",
+				Name:        "Add Reaction",
+				Description: "Add an emoji reaction to a Slack message",
+				RiskLevel:   "low",
+				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
+					"type": "object",
+					"required": ["channel", "timestamp", "name"],
+					"properties": {
+						"channel": {
+							"type": "string",
+							"description": "Channel ID containing the message (e.g. C01234567)"
+						},
+						"timestamp": {
+							"type": "string",
+							"description": "Timestamp of the message to react to (e.g. 1234567890.123456)"
+						},
+						"name": {
+							"type": "string",
+							"description": "Emoji name without colons (e.g. thumbsup, white_check_mark)"
+						}
+					}
+				}`)),
+			},
+		},
+		RequiredCredentials: []connectors.ManifestCredential{
+			{Service: "slack", AuthType: "custom", InstructionsURL: "https://api.slack.com/tutorials/tracks/getting-a-token"},
+		},
+		Templates: []connectors.ManifestTemplate{
+			{
+				ID:          "tpl_slack_send_to_channel",
+				ActionType:  "slack.send_message",
+				Name:        "Post to a channel",
+				Description: "Locks the channel and lets the agent choose the message content.",
+				Parameters:  json.RawMessage(`{"channel":"#general","message":"*"}`),
+			},
+			{
+				ID:          "tpl_slack_send_any",
+				ActionType:  "slack.send_message",
+				Name:        "Send messages freely",
+				Description: "Agent can send any message to any channel.",
+				Parameters:  json.RawMessage(`{"channel":"*","message":"*"}`),
+			},
+			{
+				ID:          "tpl_slack_create_channel",
+				ActionType:  "slack.create_channel",
+				Name:        "Create channels",
+				Description: "Agent can create public channels with any name.",
+				Parameters:  json.RawMessage(`{"name":"*","is_private":false}`),
+			},
+			{
+				ID:          "tpl_slack_list_channels",
+				ActionType:  "slack.list_channels",
+				Name:        "List channels",
+				Description: "Agent can list channels visible to the bot.",
+				Parameters:  json.RawMessage(`{"types":"*","limit":"*","cursor":"*"}`),
+			},
+			{
+				ID:          "tpl_slack_read_channel",
+				ActionType:  "slack.read_channel_messages",
+				Name:        "Read channel messages",
+				Description: "Agent can read messages from any channel.",
+				Parameters:  json.RawMessage(`{"channel":"*","limit":"*","oldest":"*","latest":"*","cursor":"*"}`),
+			},
+			{
+				ID:          "tpl_slack_read_thread",
+				ActionType:  "slack.read_thread",
+				Name:        "Read thread replies",
+				Description: "Agent can read thread replies from any channel.",
+				Parameters:  json.RawMessage(`{"channel":"*","thread_ts":"*","limit":"*","cursor":"*"}`),
+			},
+			{
+				ID:          "tpl_slack_schedule_message",
+				ActionType:  "slack.schedule_message",
+				Name:        "Schedule messages",
+				Description: "Agent can schedule messages to any channel.",
+				Parameters:  json.RawMessage(`{"channel":"*","message":"*","post_at":"*"}`),
+			},
+			{
+				ID:          "tpl_slack_set_topic",
+				ActionType:  "slack.set_topic",
+				Name:        "Set channel topics",
+				Description: "Agent can update the topic on any channel.",
+				Parameters:  json.RawMessage(`{"channel":"*","topic":"*"}`),
+			},
+			{
+				ID:          "tpl_slack_invite_to_channel",
+				ActionType:  "slack.invite_to_channel",
+				Name:        "Invite users to channels",
+				Description: "Agent can invite users to any channel.",
+				Parameters:  json.RawMessage(`{"channel":"*","users":"*"}`),
+			},
+			{
+				ID:          "tpl_slack_upload_file",
+				ActionType:  "slack.upload_file",
+				Name:        "Upload files",
+				Description: "Agent can upload files to any channel.",
+				Parameters:  json.RawMessage(`{"channel":"*","filename":"*","content":"*","title":"*"}`),
+			},
+			{
+				ID:          "tpl_slack_add_reaction",
+				ActionType:  "slack.add_reaction",
+				Name:        "Add reactions",
+				Description: "Agent can add emoji reactions to messages.",
+				Parameters:  json.RawMessage(`{"channel":"*","timestamp":"*","name":"*"}`),
+			},
+		},
+	}
+}
